+++ packages/blade/src/components/Icons/UsersIcon/UsersIcon.web.test.tsx
@@ -1,0 +1,11 @@
+import UsersIcon from '.';
+import renderWithTheme from '~src/_helpers/testing/renderWithTheme.web';
+
+describe('<UsersIcon />', () => {
+  it('should render UsersIcon', () => {
+    const { container } = renderWithTheme(
+      <UsersIcon color="feedback.icon.neutral.lowContrast" size="large" />,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});