+++ packages/blade/src/components/Indicator/Indicator.tsx
@@ -7,7 +7,7 @@ import Circle from '~components/Icons/_Svg/Circle';
 import { Text } from '~components/Typography';
 
 import type { Feedback } from '~tokens/theme/theme';
-import { getPlatformType, makeAccessible } from '~utils';
+import { metaAttribute, getPlatformType, makeAccessible, MetaConstants } from '~utils';
 
 type IndicatorCommonProps = {
   /**
@@ -86,7 +86,13 @@ const Indicator = ({
   });
 
   return (
-    <Box display="flex" flexDirection="row" alignItems="center" {...a11yProps}>
+    <Box
+      display="flex"
+      flexDirection="row"
+      alignItems="center"
+      {...a11yProps}
+      {...metaAttribute(MetaConstants.Component, MetaConstants.Indicator)}
+    >
       <Svg width={dimensions.svgSize} height={dimensions.svgSize} viewBox="0 0 10 10" fill="none">
         <Circle cx="5" cy="5" r="5" fill={fillColor} />
         <Circle cx="5" cy="5" r="4.75" stroke={strokeColor} strokeWidth="0.5" />