+++ packages/blade/src/components/Input/TextInput/TextInput.tsx
@@ -4,11 +4,12 @@ import type { TextInput as TextInputReactNative } from 'react-native';
 import type { BaseInputProps } from '../BaseInput';
 import { BaseInput } from '../BaseInput';
 import type { IconComponent } from '~components/Icons';
-import { InfoIcon, CloseIcon } from '~components/Icons';
+import { CloseIcon } from '~components/Icons';
 import { IconButton } from '~components/Button/IconButton';
 import { getPlatformType, isEmpty } from '~utils';
 import { CharacterCounter } from '~components/Form/CharacterCounter';
 import Box from '~components/Box';
+import { Spinner } from '~components/Spinner';
 
 // Users should use PasswordInput for input type password
 type Type = Exclude<BaseInputProps['type'], 'password'>;
@@ -185,8 +186,7 @@ export const TextInput = ({
 
   const renderInteractionElement = (): ReactNode => {
     if (isLoading) {
-      // @TODO replace with spinner once it's ready
-      return <InfoIcon size="medium" color="surface.text.subtle.lowContrast" />;
+      return <Spinner accessibilityLabel="Loading Content" />;
     }
 
     if (shouldShowClearButton) {
@@ -221,6 +221,7 @@ export const TextInput = ({
   return (
     <BaseInput
       id="textinput"
+      componentName="textinput"
       ref={textInputRef as React.Ref<HTMLInputElement>}
       label={label}
       labelPosition={labelPosition}