+++ packages/blade/src/components/Spinner/Spinner/__tests__/Spinner.web.test.tsx
@@ -3,32 +3,32 @@ import renderWithTheme from '~src/_helpers/testing/renderWithTheme.web';
 
 describe('<Spinner />', () => {
   it('should render Spinner with default props', () => {
-    const { container } = renderWithTheme(<Spinner />);
-    expect(container).toMatchSnapshot();
-  });
-
-  it('should render small size Spinner', () => {
-    const { container } = renderWithTheme(<Spinner size="small" />);
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" />);
     expect(container).toMatchSnapshot();
   });
 
   it('should render medium size Spinner', () => {
-    const { container } = renderWithTheme(<Spinner size="medium" />);
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" size="medium" />);
     expect(container).toMatchSnapshot();
   });
 
   it('should render large size Spinner', () => {
-    const { container } = renderWithTheme(<Spinner size="large" />);
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" size="large" />);
+    expect(container).toMatchSnapshot();
+  });
+
+  it('should render xlarge size Spinner', () => {
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" size="xlarge" />);
     expect(container).toMatchSnapshot();
   });
 
   it('should render low contrast Spinner', () => {
-    const { container } = renderWithTheme(<Spinner contrast="low" />);
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" contrast="low" />);
     expect(container).toMatchSnapshot();
   });
 
   it('should render high contrast Spinner', () => {
-    const { container } = renderWithTheme(<Spinner contrast="high" />);
+    const { container } = renderWithTheme(<Spinner accessibilityLabel="Loading" contrast="high" />);
     expect(container).toMatchSnapshot();
   });
 });