+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @razorpay/blade
 
+## 5.1.2
+
+### Patch Changes
+
+- bd0b675: chore(blade): added blade component data attributes
+
+## 5.1.1
+
+### Patch Changes
+
+- 5a6b980: feat: add Mail icon
+
+## 5.1.0
+
+### Minor Changes
+
+- d4b981e: fix: show `Spinner` on `TextInput` when `isLoading=true`
+  - Adds spinner when `isLoading: true` is passed to `TextInput`. This was a long pending TODO
+  - Update Spinner sizes after the design was updated \* This doesn't need any code mod since there are 9 instances of spinner being used with default variant i.e medium
+
+## 5.0.1
+
+### Patch Changes
+
+- 96cf25f: feat: add new icons (lock, settings, file-text, users, slash)
+
 ## 5.0.0
 
 ### Major Changes